+++ tests/Schema/Schema.Test.js
@@ -2675,5 +2675,226 @@ describe('Schema', () => {
 			assert.equal(testVar, 4);
 		});
 	});
+
+	describe('.extend', () => {
+		it('should accept undefined', () => {
+			let total = 0;
+			let testVar = 0;
+
+			const schema = new Schema({
+				testKey: [{
+					level2: Number
+				}],
+				testKey2: String,
+				testKey3: Date
+			});
+
+			const result = schema.extend();
+
+			schema.eachRule((path, rule) => {
+				total++;
+				if (deepEqual(path, []) && rule.types[0].type === Object) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey']) && rule.types[0].type === Array) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey', 0]) && rule.types[0].type === Object) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey', 0, 'level2']) && rule.types[0].type === Number) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey2']) && rule.types[0].type === String) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey3']) && rule.types[0].type === Date) {
+					testVar++;
+				}
+			});
+
+			result.eachRule((path, rule) => {
+				total++;
+				if (deepEqual(path, []) && rule.types[0].type === Object) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey']) && rule.types[0].type === Array) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey', 0]) && rule.types[0].type === Object) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey', 0, 'level2']) && rule.types[0].type === Number) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey2']) && rule.types[0].type === String) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey3']) && rule.types[0].type === Date) {
+					testVar++;
+				}
+			});
+
+			assert.equal(total, 12);
+			assert.equal(testVar, 12);
+			assert.notEqual(schema, result);
+		});
+
+		it('should accept an object', () => {
+			let total = 0;
+			let testVar = 0;
+
+			const schema = new Schema({
+				testKey: [{
+					level2: Number
+				}],
+				testKey2: String,
+				testKey3: Date
+			});
+
+			const result = schema.extend({
+				testKey: [{
+					level2: 'integer'
+				}],
+				testKey2: Number
+			});
+
+			schema.eachRule((path, rule) => {
+				total++;
+				if (deepEqual(path, []) && rule.types[0].type === Object) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey']) && rule.types[0].type === Array) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey', 0]) && rule.types[0].type === Object) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey', 0, 'level2']) && rule.types[0].type === Number) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey2']) && rule.types[0].type === String) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey3']) && rule.types[0].type === Date) {
+					testVar++;
+				}
+			});
+
+			result.eachRule((path, rule) => {
+				total++;
+				if (deepEqual(path, []) && rule.types[0].type === Object) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey']) && rule.types[0].type === Array) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey', 0]) && rule.types[0].type === Object) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey', 0, 'level2']) && rule.types[0].type === 'integer') {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey2']) && rule.types[0].type === Number) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey3']) && rule.types[0].type === Date) {
+					testVar++;
+				}
+			});
+
+			assert.equal(total, 12);
+			assert.equal(testVar, 12);
+		});
+
+		it('should accept an instance of Schema', () => {
+			let total = 0;
+			let testVar = 0;
+
+			const schema = new Schema({
+				testKey: [{
+					level2: Number
+				}],
+				testKey2: String,
+				testKey3: Date
+			});
+			const schema2 = new Schema({
+				testKey: [{
+					level2: 'integer'
+				}],
+				testKey2: Number
+			});
+
+			const result = schema.extend(schema2);
+
+			schema.eachRule((path, rule) => {
+				total++;
+				if (deepEqual(path, []) && rule.types[0].type === Object) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey']) && rule.types[0].type === Array) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey', 0]) && rule.types[0].type === Object) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey', 0, 'level2']) && rule.types[0].type === Number) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey2']) && rule.types[0].type === String) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey3']) && rule.types[0].type === Date) {
+					testVar++;
+				}
+			});
+
+			schema2.eachRule((path, rule) => {
+				total++;
+				if (deepEqual(path, []) && rule.types[0].type === Object) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey']) && rule.types[0].type === Array) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey', 0]) && rule.types[0].type === Object) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey', 0, 'level2']) && rule.types[0].type === 'integer') {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey2']) && rule.types[0].type === Number) {
+					testVar++;
+				}
+			});
+
+			result.eachRule((path, rule) => {
+				total++;
+				if (deepEqual(path, []) && rule.types[0].type === Object) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey']) && rule.types[0].type === Array) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey', 0]) && rule.types[0].type === Object) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey', 0, 'level2']) && rule.types[0].type === 'integer') {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey2']) && rule.types[0].type === Number) {
+					testVar++;
+				}
+				if (deepEqual(path, ['testKey3']) && rule.types[0].type === Date) {
+					testVar++;
+				}
+			});
+
+			assert.equal(total, 17);
+			assert.equal(testVar, 17);
+		});
+
+		// TODO: should accept undefined
+	});
 });