+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## Unrelesed
+### Added
+- Method 'extend' in [Schema](docs/Schema.md)
+- Method 'extend' in [Model](docs/Model.md)
+
 ## [0.4.0] - 2019-07-01
 ### Added
 - Indexing in [Collection](docs/Collection.md)